Keep AI stream alive across page navigation
